아이템매니저데이터넣어야해

+++ princessMaker2/item.cpp
@@ -29,7 +29,7 @@ void item::release()
 {
 }
 
-void item::setItem(string name, int price, vector<pair<string, int>> property, int type, int frameX, int frameY)
+void item::setItem(string name, int price, vector<pair<string, float>> property, int type, int frameX, int frameY)
 {
 	_name = name;
 	_price = price;
