+++ princessMaker2/item.h
@@ -10,9 +10,10 @@ private:
 	ITEM_TYPE _type;
 	int _frameX, _frameY;
 	image* _img;
-	vector<pair<string, int>> _property;
+	vector<pair<string, float>> _property;
 	bool _isWear;
 	float _x, _y;
+
 public:
 	item();
 	~item();
@@ -22,7 +23,7 @@ public:
 	void render();
 	void release();
 
-	void setItem(string name, int price, vector<pair<string, int>> property, int type, int frameX, int frameY);
+	void setItem(string name, int price, vector<pair<string, float>> property, int type, int frameX, int frameY);
 	void setImage();
 	void setXY(float x, float y);
 	float getX() { return _x; }
@@ -30,6 +31,6 @@ public:
 
 	string getName() { return _name; }
 	int getPrice() { return _price; }
-	vector<pair<string,int>> getProperty() { return _property; }
+	vector<pair<string, float>> getProperty() { return _property; }
 };
